Fix/issue 78 size t conversion (#93)
* Fixes and optimizations from code review

- Fix nec_ground::get_radial_wire_length() returning wrong field
- Fix dangling pointer in ASSERT macro (nec_debug.h)
- Remove using namespace std from nec2cpp.h header
- Fix multi-frequency plane wave excitation loop
- Optimize safe_array::resize() to use geometric growth
- Refactor c_ggrid::interpolate() for thread safety and speed
- Fix safe_array BoundsViol ODR issue

* Fix #78: Add static_cast<size_t> to memcpy in safe_array::resize()

On 32-bit platforms, int64_t is wider than size_t, causing a
-Wconversion warning when int64_t _len is passed to std::memcpy.
Explicitly cast to avoid the warning.

Closes #78

+++ src/safe_array.h
@@ -117,7 +117,7 @@ public:
       try {
         T* new_data_ = new T[_data_size];
         if (0 != _len)
-          std::memcpy(new_data_, _data, _len * sizeof(T));
+          std::memcpy(new_data_, _data, static_cast<size_t>(_len) * sizeof(T));
 
         delete[] _data;
         _data = new_data_;
